create separate class for database constraint tests
